refactor(NewsletterSchema): improve file naming
